temporary bug fix in cd._compute_kstar: casting distance type to float64
distances types in cython functions must be float64, but this is never checked in the cython functions in density_estimation!

+++ dadapy/density_estimation.py
@@ -194,7 +194,7 @@ class DensityEstimation(IdEstimation):
             self.maxk,
             Dthr,
             self.dist_indices,
-            self.distances,
+            self.distances.astype('float64'),
         )
         self.set_kstar(kstar)
